Pipeline pods volumes (#8)
Bottom DAG nodes extracted, passed to DAG template. 2 tasks added, before the start of the pipeline to create volume and when all bottom nodes are done, to delete the volume. Using templated pod definition so that uniquely generated volume can be injected to pod definition. Volume size and policy are configurable.

Co-authored-by: Mariusz Strzelecki <mariusz.strzelecki@getindata.com>